requirements.txt
flask==2.2.3
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==1.0.0
pyjwt==2.6.0
bcrypt==4.0.1
python-dateutil==2.8.2